split Style.go into StyleSetter.go - add more comments
